check earlier for attributes that are placed in `where` predicate
